pyproject.toml
[project]
name = "fund-rank"
version = "0.1.0"
description = "Brazilian fixed-income fund ranking pipeline (Decade PS take-home)"
authors = [{name = "pedroamsdo"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "polars>=1.18.0",
    "duckdb>=1.1.0",
    "pyarrow>=18.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.28.0",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "fsspec>=2024.10.0",
    "typer>=0.15.0",
    "pyyaml>=6.0.2",
    "python-dateutil>=2.9.0",
    "eval-type-backport>=0.2.0;python_version<'3.10'",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
]
prefect = [
    "prefect>=3.1.0",
]

[project.scripts]
fund-rank = "fund_rank.cli:app"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fund_rank"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --strict-markers"
markers = [
    "integration: integration tests requiring network or large fixtures",
]
